+++ config/routes.rb
@@ -62,6 +62,7 @@ Pwb::Engine.routes.draw do
     namespace :import do
       get "/mls_experiment" => "mls#experiment"
       get "/mls" => "mls#retrieve"
+      get "/scrapper" => "scrapper#retrieve"
       post "/properties/retrieve_from_pwb" => "properties#retrieve_from_pwb"
       post "/properties/retrieve_from_mls" => "properties#retrieve_from_mls"
       post "/translations" => "translations#multiple"